requirements.txt
-i https://pypi.org/simple
aiosqlite==0.17.0 ; python_version >= '3.6'
alembic==1.9.0 ; python_version >= '3.7'
anyio==3.6.2 ; python_full_version >= '3.6.2'
apache-airflow==2.5.0
apache-airflow-providers-common-sql==1.3.1 ; python_version ~= '3.7'
apache-airflow-providers-ftp==3.2.0 ; python_version ~= '3.7'
apache-airflow-providers-http==4.1.0 ; python_version ~= '3.7'
apache-airflow-providers-imap==3.1.0 ; python_version ~= '3.7'
apache-airflow-providers-sqlite==3.3.1 ; python_version ~= '3.7'
apispec[yaml]==3.3.2 ; python_version >= '3.5'
apprise==1.2.0 ; python_version >= '3.6'
argcomplete==2.0.0 ; python_version >= '3.6'
asgi-lifespan==2.0.0 ; python_version >= '3.7'
asyncpg==0.27.0 ; python_full_version >= '3.7.0'
attrs==22.1.0 ; python_version >= '3.5'
babel==2.11.0 ; python_version >= '3.6'
blinker==1.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
boto3==1.26.32
botocore==1.29.32 ; python_version >= '3.7'
cachelib==0.9.0 ; python_version >= '3.7'
cachetools==5.2.0 ; python_version ~= '3.7'
cattrs==22.2.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
clickclick==20.10.2
cloudpickle==2.2.0 ; python_version >= '3.6'
colorama==0.4.6 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
colorlog==4.8.0
commonmark==0.9.1
configupdater==3.1.1 ; python_version >= '3.6'
connexion[flask,swagger-ui]==2.14.1 ; python_version >= '3.6'
coolname==2.1.0
cron-descriptor==1.2.32
croniter==1.3.8 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
cryptography==38.0.4 ; python_version >= '3.6'
deprecated==1.2.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dill==0.3.6 ; python_version >= '3.7'
dnspython==2.2.1 ; python_version >= '3.6' and python_version < '4.0'
docker==6.0.1 ; python_version >= '3.7'
docutils==0.19 ; python_version >= '3.7'
email-validator==1.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
exceptiongroup==1.0.4 ; python_version < '3.11'
fastapi==0.88.0 ; python_version >= '3.7'
flask==2.2.2 ; python_version >= '3.7'
flask-appbuilder==4.1.4 ; python_version ~= '3.7'
flask-babel==2.0.0
flask-caching==2.0.1 ; python_version >= '3.7'
flask-jwt-extended==4.4.4 ; python_version >= '3.7' and python_version < '4'
flask-login==0.6.2 ; python_version >= '3.7'
flask-session==0.4.0
flask-sqlalchemy==2.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-wtf==1.0.1 ; python_version >= '3.6'
fsspec==2022.11.0 ; python_version >= '3.7'
google-auth==2.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
graphviz==0.20.1 ; python_version >= '3.7'
greenlet==2.0.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
griffe==0.25.0 ; python_version >= '3.7'
gunicorn==20.1.0 ; python_version >= '3.5'
h11==0.14.0 ; python_version >= '3.7'
h2==4.1.0
hpack==4.0.0 ; python_full_version >= '3.6.1'
httpcore==0.16.2 ; python_version >= '3.7'
httpx==0.23.1 ; python_version >= '3.7'
hyperframe==6.0.1 ; python_full_version >= '3.6.1'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==5.1.0 ; python_version >= '3.7'
inflection==0.5.1 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
jsonpatch==1.32 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
jsonpointer==2.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonschema==4.17.3 ; python_version >= '3.7'
keyring==23.4.1 ; python_version >= '3.6'
kubernetes==25.3.0 ; python_version >= '3.6'
lazy-object-proxy==1.8.0 ; python_version >= '3.7'
linkify-it-py==2.0.0 ; python_version >= '3.6'
lockfile==0.12.2
mako==1.2.4 ; python_version >= '3.7'
markdown==3.4.1 ; python_version >= '3.7'
markdown-it-py==2.1.0 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
marshmallow==3.19.0 ; python_version >= '3.7'
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1 ; python_version >= '3.6'
marshmallow-sqlalchemy==0.26.1 ; python_version >= '3.6'
mdit-py-plugins==0.3.3 ; python_version >= '3.7'
mdurl==0.1.2 ; python_version >= '3.7'
numpy==1.23.5 ; python_version >= '3.10'
oauthlib==3.2.2 ; python_version >= '3.6'
orjson==3.8.3 ; python_version >= '3.7'
packaging==22.0 ; python_version >= '3.7'
pandas==1.5.2 ; python_version >= '3.8'
pathspec==0.9.0
pendulum==2.1.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pluggy==1.0.0 ; python_version >= '3.6'
prefect==2.7.3 ; python_version >= '3.7'
prison==0.2.1
psutil==5.9.4 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2 ; python_version >= '3.7'
pygments==2.13.0 ; python_version >= '3.6'
pyjwt==2.6.0 ; python_version >= '3.7'
pyrsistent==0.19.2 ; python_version >= '3.7'
python-daemon==2.3.2 ; python_version >= '3'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-nvd3==0.15.0
python-slugify==7.0.0 ; python_version >= '3.7'
pytz==2022.6
pytzdata==2020.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0 ; python_version >= '3.6'
readchar==4.0.3 ; python_version >= '3.6'
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests-toolbelt==0.10.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3986[idna2008]==1.5.0
rich==12.6.0 ; python_full_version >= '3.6.3' and python_full_version < '4.0.0'
rsa==4.9 ; python_version >= '3.6'
s3transfer==0.6.0 ; python_version >= '3.7'
sagetasks==0.4.0
setproctitle==1.3.2 ; python_version >= '3.7'
setuptools==65.6.3 ; python_version >= '3.7'
sevenbridges-python==2.9.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==1.4.45 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy-jsonfield==1.0.0 ; python_full_version >= '3.5.0'
sqlalchemy-utils==0.38.3 ; python_version ~= '3.6'
sqlparse==0.4.3 ; python_version >= '3.5'
starlette==0.22.0 ; python_version >= '3.7'
swagger-ui-bundle==0.0.9
synapseclient==2.7.0 ; python_version >= '3.7'
tabulate==0.9.0 ; python_version >= '3.7'
tenacity==8.1.0 ; python_version >= '3.6'
termcolor==2.1.1 ; python_version >= '3.7'
text-unidecode==1.3
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typer==0.7.0 ; python_version >= '3.6'
typing-extensions==4.4.0 ; python_version >= '3.7'
uc-micro-py==1.0.1 ; python_version >= '3.6'
unicodecsv==0.14.1
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.20.0 ; python_version >= '3.7'
websocket-client==1.4.2 ; python_version >= '3.7'
werkzeug==2.2.2 ; python_version >= '3.7'
wrapt==1.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wtforms==3.0.1 ; python_version >= '3.7'
zipp==3.11.0 ; python_version >= '3.7'
